Add CodesSolver base class
This is intended to replace the FileProgramInterface class, as part of
the refactor of the solvers interface.
